requirements.txt
streamlit==1.32.0
requests
python-dotenv
pypdf